test(cli): extract unit target play tests
Move the unit-target command and postcondition tests into a focused play suite. The new file owns the readUnitTargetAction fake, delayed/no-op/path-shortfall postcondition cases, and GamePlayUnitTarget assertions.

Keep the monolith responsible for read-side ready-unit, unit-move-preview, notification, and compact-priority projection fixtures. Wire the focused suite into test:cli:play so the extracted coverage remains canonical.

Proof: git diff --check; bun run --cwd packages/cli test -- game.play.unit-target.test.ts; bun run --cwd packages/cli test -- game.play.test.ts -t "unit target|unit-target|unit move|ready-unit|compact priorities"; bun run check:cli; bun run test:cli:play. This is test ownership modularization only; it changes no runtime behavior, direct-control behavior, live-game flow, gameplay send, play-thread workflow, Studio/oRPC transport, or relationship/city-state/suzerain evidence policy.
