+++ packages/cli/test/commands/game.play.test.ts
@@ -22,7 +22,6 @@ import GamePlayTargetCandidates from '../../src/commands/game/play/target-candid
 import GamePlayTopics from '../../src/commands/game/play/topics';
 import GamePlayTraditions from '../../src/commands/game/play/traditions';
 import GamePlayUnitMovePreview from '../../src/commands/game/play/unit-move-preview';
-import GamePlayUnitTarget from '../../src/commands/game/play/unit-target';
 import GameWatch from '../../src/commands/game/watch';
 import { startFakeTunerServer } from './fixtures/tuner-socket-server';
 
@@ -1960,168 +1959,6 @@ describe('game play commands', () => {
     }
   });
 
-  test('resolves unit target actions without sending by default', async () => {
-    const server = await startTunerServer();
-    try {
-      const { port } = server.address();
-      await GamePlayUnitTarget.run([
-        '--host',
-        '127.0.0.1',
-        '--port',
-        String(port),
-        '--unit-id',
-        '{"owner":0,"id":65536,"type":26}',
-        '--x',
-        '23',
-        '--y',
-        '33',
-        '--json',
-      ]);
-
-      expect(server.received.some((message) => message.includes('readUnitTargetAction'))).toBe(true);
-      expect(server.received.some((message) => message.includes('operationType.replace(/^UNITOPERATION_/'))).toBe(true);
-      expect(server.received.some((message) => message.includes('"send":true'))).toBe(false);
-    } finally {
-      await server.close();
-    }
-  });
-
-  test('surfaces sent unit-target no-ops as postcondition misses', async () => {
-    const server = await startTunerServer({ unitTargetMode: 'no-op-after-send' });
-    const writes: string[] = [];
-    const log = vi.spyOn(GamePlayUnitTarget.prototype, 'log').mockImplementation((message?: string) => {
-      if (message) writes.push(message);
-    });
-    try {
-      const { port } = server.address();
-      await GamePlayUnitTarget.run([
-        '--host',
-        '127.0.0.1',
-        '--port',
-        String(port),
-        '--unit-id',
-        '{"owner":0,"id":65536,"type":26}',
-        '--x',
-        '23',
-        '--y',
-        '33',
-        '--send',
-        '--reason',
-        'test postcondition miss',
-        '--json',
-      ]);
-
-      const payload = JSON.parse(writes.join('')) as {
-        ok: true;
-        result: { sent: boolean; verified: boolean; verification: { status: string; classification: string; reason: string } };
-      };
-      expect(payload.result.sent).toBe(true);
-      expect(payload.result.verified).toBe(false);
-      expect(payload.result.verification.status).toBe('no-state-change');
-      expect(payload.result.verification.classification).toBe('no-state-change');
-      expect(payload.result.verification.reason).toMatch(/re-read .*before repeating/);
-      expect(server.received.some((message) => message.includes('"send":true'))).toBe(true);
-    } finally {
-      log.mockRestore();
-      await server.close();
-    }
-  });
-
-  test('stabilizes delayed unit-target postconditions before returning', async () => {
-    const server = await startTunerServer({ unitTargetMode: 'delayed-after-send' });
-    const writes: string[] = [];
-    const log = vi.spyOn(GamePlayUnitTarget.prototype, 'log').mockImplementation((message?: string) => {
-      if (message) writes.push(message);
-    });
-    try {
-      const { port } = server.address();
-      await GamePlayUnitTarget.run([
-        '--host',
-        '127.0.0.1',
-        '--port',
-        String(port),
-        '--unit-id',
-        '{"owner":0,"id":65536,"type":26}',
-        '--x',
-        '23',
-        '--y',
-        '33',
-        '--send',
-        '--reason',
-        'test delayed postcondition',
-        '--json',
-      ]);
-
-      const payload = JSON.parse(writes.join('')) as {
-        ok: true;
-        result: { sent: boolean; verified: boolean; verification: { status: string; classification: string; source: string; attempts: number; reason: string } };
-      };
-      expect(payload.result.sent).toBe(true);
-      expect(payload.result.verified).toBe(true);
-      expect(payload.result.verification.status).toBe('verified');
-      expect(payload.result.verification.classification).toBe('unit-state-changed');
-      expect(payload.result.verification.source).toBe('bounded-poll');
-      expect(payload.result.verification.attempts).toBeGreaterThan(0);
-      expect(payload.result.verification.reason).toMatch(/bounded post-send polling/);
-      expect(server.received.filter((message) => message.includes('readUnitTargetAction')).length).toBeGreaterThan(1);
-    } finally {
-      log.mockRestore();
-      await server.close();
-    }
-  });
-
-  test('classifies sent MOVE_TO short landings as path shortfalls', async () => {
-    const server = await startTunerServer({ unitTargetMode: 'path-shortfall' });
-    const writes: string[] = [];
-    const log = vi.spyOn(GamePlayUnitTarget.prototype, 'log').mockImplementation((message?: string) => {
-      if (message) writes.push(message);
-    });
-    try {
-      const { port } = server.address();
-      await GamePlayUnitTarget.run([
-        '--host',
-        '127.0.0.1',
-        '--port',
-        String(port),
-        '--unit-id',
-        '{"owner":0,"id":65536,"type":26}',
-        '--x',
-        '23',
-        '--y',
-        '33',
-        '--send',
-        '--reason',
-        'test movement path shortfall',
-        '--json',
-      ]);
-
-      const payload = JSON.parse(writes.join('')) as {
-        ok: true;
-        result: {
-          verified: boolean;
-          verification: {
-            status: string;
-            classification: string;
-            destinationReached: boolean;
-            requestedLocation: { x: number; y: number };
-            landedLocation: { x: number; y: number };
-            reason: string;
-          };
-        };
-      };
-      expect(payload.result.verified).toBe(true);
-      expect(payload.result.verification.status).toBe('verified');
-      expect(payload.result.verification.classification).toBe('path-shortfall');
-      expect(payload.result.verification.destinationReached).toBe(false);
-      expect(payload.result.verification.requestedLocation).toEqual({ x: 23, y: 33 });
-      expect(payload.result.verification.landedLocation).toEqual({ x: 22, y: 34 });
-      expect(payload.result.verification.reason).toMatch(/landed short/);
-    } finally {
-      log.mockRestore();
-      await server.close();
-    }
-  });
-
   test('reads ready-unit tactical view without sending operations', async () => {
     const server = await startTunerServer();
     try {
@@ -2846,10 +2683,8 @@ function expectOwnerOnlyContactLabels(values: readonly string[]): void {
 
 async function startTunerServer(options: {
   playNotificationMode?: 'town-focus' | 'production-choice' | 'population-placement' | 'tech-choice' | 'culture-choice' | 'celebration-choice' | 'government-choice' | 'narrative-choice' | 'narrative-choice-empty' | 'narrative-choice-visible-panel' | 'tradition-review' | 'stale-unit-command' | 'stale-unit-command-disabled' | 'stale-unit-command-pending' | 'stale-informational' | 'unit-lost-report' | 'legacy-completed' | 'diplomatic-report' | 'diplomatic-action-report' | 'first-meet' | 'ready-unit' | 'mixed-queue' | 'clean-read' | 'stale-diplomacy' | 'runtime-error';
-  unitTargetMode?: 'verified' | 'no-op-after-send' | 'path-shortfall' | 'delayed-after-send';
   notificationDismissalMode?: 'verified' | 'stale-nonblocking' | 'engine-front-train-absent' | 'engine-front-dismissed';
 } = {}) {
-  let unitTargetSendObserved = false;
   let notificationDismissalSent = false;
   return startFakeTunerServer({
     handle({ message }) {
@@ -2863,14 +2698,6 @@ async function startTunerServer(options: {
           key: 'PLAYER_REALATIONSHIP_FIRSTMEET_NEUTRAL',
           value: 673478009,
         })];
-      }
-      if (message.includes('readUnitTargetAction')) {
-        const send = message.includes('"send":true');
-        if (send) unitTargetSendObserved = true;
-        const mode = options.unitTargetMode === 'delayed-after-send' && unitTargetSendObserved && !send
-          ? 'delayed-observed'
-          : options.unitTargetMode;
-        return [JSON.stringify(unitTargetAction(send, mode))];
       }
       if (message.includes('readUnitMovePreview')) {
         return [JSON.stringify(unitMovePreviewView())];
@@ -5117,86 +4944,6 @@ function playNotificationView(
       ],
     },
     limits: { maxNotifications: 25, truncated: false },
-  };
-}
-
-function unitTargetAction(send: boolean, mode: 'verified' | 'no-op-after-send' | 'path-shortfall' | 'delayed-after-send' | 'delayed-observed' = 'verified') {
-  const unitId = { owner: 0, id: 65536, type: 26 };
-  const beforeUnit = { ok: true, value: { id: unitId, location: { x: 22, y: 33 }, movementMovesRemaining: 2, attacksRemaining: 1 } };
-  const delayedObservedUnit = { ok: true, value: { id: unitId, location: { x: 22, y: 33 }, movementMovesRemaining: 2, attacksRemaining: 0 } };
-  const beforeTargetUnits = { ok: true, value: [{ owner: 62, id: 123, type: 26 }] };
-  const verified = send && mode === 'verified';
-  const pathShortfall = send && mode === 'path-shortfall';
-  const delayedObserved = !send && mode === 'delayed-observed';
-  const selected = mode === 'path-shortfall'
-    ? {
-        family: 'unit-operation',
-        operationType: 'MOVE_TO',
-        args: { X: 23, Y: 33, Modifiers: 3 },
-        valid: true,
-        result: { Success: true, Plots: [1457] },
-        targetInReturnedPlots: true,
-      }
-    : {
-        family: 'unit-operation',
-        operationType: 'UNITOPERATION_RANGE_ATTACK',
-        args: { X: 23, Y: 33, Modifiers: 3 },
-        valid: true,
-        result: { Success: true, Plots: [1457] },
-        targetInReturnedPlots: true,
-      };
-  return {
-    unitId,
-    target: { x: 23, y: 33, index: { ok: true, value: 1457 } },
-    beforeUnit: delayedObserved ? delayedObservedUnit : beforeUnit,
-    beforeTargetUnits,
-    candidates: [selected],
-    selected,
-    sent: send,
-    ...(send
-      ? {
-          sendResult: { accepted: true },
-          afterUnit: verified || pathShortfall
-            ? {
-                ok: true,
-                value: {
-                  id: unitId,
-                  location: pathShortfall ? { x: 22, y: 34 } : { x: 22, y: 33 },
-                  movementMovesRemaining: pathShortfall ? 0 : 2,
-                  attacksRemaining: verified ? 0 : 1,
-                },
-              }
-            : beforeUnit,
-          afterTargetUnits: beforeTargetUnits,
-          verified: verified || pathShortfall,
-          verification: {
-            status: verified || pathShortfall ? 'verified' : 'no-state-change',
-            classification: pathShortfall ? 'path-shortfall' : verified ? 'unit-state-changed' : 'no-state-change',
-            unitChanged: verified || pathShortfall,
-            targetUnitsChanged: false,
-            destinationReached: pathShortfall ? false : null,
-            requestedLocation: { x: 23, y: 33 },
-            landedLocation: pathShortfall ? { x: 22, y: 34 } : { x: 22, y: 33 },
-            reason: pathShortfall
-              ? 'unit moved, but landed short of the requested target tile; re-read before issuing a follow-up move'
-              : verified
-                ? 'unit state changed after send'
-              : 'send returned but unit and target-plot probes did not change; re-read before repeating',
-          },
-        }
-      : {
-          verification: {
-            status: 'not-sent',
-            classification: 'not-sent',
-            unitChanged: false,
-            targetUnitsChanged: false,
-            destinationReached: null,
-            requestedLocation: { x: 23, y: 33 },
-            landedLocation: { x: 22, y: 33 },
-            reason: 'read-only target resolution; use --send with an approval reason to mutate',
-          },
-        }),
-    notes: ['Selection follows the official right-click WorldInput target order.'],
   };
 }
